PASN: Extend RSNXE capability field from u32 to u64
wpa_pasn_add_rsnxe(), pasn_set_rsnxe_caps(), and the rsnxe_capab field
in struct pasn_data all used u32, limiting RSNXE capability
representation to 32 bits (4 octets). WLAN_RSNX_CAPAB_SAE_PW_ID_CHANGE
is bit 34, which requires at least 5 octets.

Widen the type to u64 throughout and extend wpa_pasn_add_rsnxe() to
write up to 8 octets of the little-endian capability field, matching the
approach already used in wpa_ie.c (BIT_ULL).

+++ src/pasn/pasn_common.c
@@ -175,7 +175,7 @@ void pasn_set_rsn_pairwise(struct pasn_data *pasn, int rsn_pairwise)
 }
 
 
-void pasn_set_rsnxe_caps(struct pasn_data *pasn, u32 rsnxe_capab)
+void pasn_set_rsnxe_caps(struct pasn_data *pasn, u64 rsnxe_capab)
 {
 	if (!pasn)
 		return;
